+++ src/main/scala/org/podval/tools/testing/task/TestTask.scala
@@ -2,24 +2,24 @@ package org.podval.tools.testing.task
 
 import groovy.lang.{Closure, DelegatesTo}
 import org.gradle.StartParameter
-import org.gradle.api.file.FileCollection
 import org.gradle.api.internal.DocumentationRegistry
 import org.gradle.api.internal.tasks.testing.filter.DefaultTestFilter
 import org.gradle.api.plugins.JavaBasePlugin
 import org.gradle.api.plugins.scala.ScalaBasePlugin
 import org.gradle.api.tasks.testing.{AbstractTestTask, Test, TestListener}
-import org.gradle.api.tasks.{Classpath, SourceSet, TaskAction}
-import org.gradle.api.{Action, Project}
+import org.gradle.api.tasks.{SourceSet, TaskAction}
+import org.gradle.api.Action
 import org.gradle.internal.event.ListenerBroadcast
 import org.gradle.internal.time.Clock
 import org.gradle.internal.work.WorkerLeaseService
 import org.gradle.internal.{Actions, Cast}
 import org.gradle.util.internal.ConfigureUtil
 import org.opentorah.build.Gradle.*
+import org.opentorah.build.{GradleClassPath, TaskWithSourceSet}
 import org.opentorah.util.Files
 import java.io.File
 import java.lang.reflect.{Field, Method}
-import scala.jdk.CollectionConverters.*
+import scala.jdk.CollectionConverters.IterableHasAsScala
 
 // guide: https://docs.gradle.org/current/userguide/java_testing.html
 // configuration: https://docs.gradle.org/current/dsl/org.gradle.api.tasks.testing.Test.html
@@ -27,8 +27,15 @@ import scala.jdk.CollectionConverters.*
 // Note: inherited Test.testsAreNotFiltered() calls Test.noCategoryOrTagOrGroupSpecified(),
 // which recognizes only JUnit and TestNG; since I can not override it, I'll just use
 // org.gradle.api.tasks.testing.junit.JUnitOptions for my "sbt" TestFrameworkOptions...
-abstract class TestTask extends Test:
+abstract class TestTask extends Test with TaskWithSourceSet:
+  // To avoid invoking Task.getProject at execution time, some things are done or captured at create or right before execution:
   setGroup(JavaBasePlugin.VERIFICATION_GROUP)
+  getDependsOn.add(getProject.getClassesTask(sourceSet))
+
+  private val buildDirectory: File = getProject.getLayout.getBuildDirectory.get.getAsFile
+  private val scalaCompile: String = getProject.getScalaCompile(sourceSet).getName
+  private val zincClassPath: Iterable[File] = getProject.getConfiguration(ScalaBasePlugin.ZINC_CONFIGURATION_NAME).asScala
+  private var analysisFile: Option[File] = None
 
   // Note: Deferring creation of the test framework by using
   // `getTestFrameworkProperty.convention(project.provider(() => createTestFramework))` did not work out,
@@ -45,26 +52,7 @@ abstract class TestTask extends Test:
   //   at org.podval.tools.testing.task.TestTaskScala_Decorated.<init>(Unknown Source)
   useSbt()
 
-  // TODO unify with ScalaJSTask?
-  private def sourceSet: SourceSet = getProject.getSourceSet(SourceSet.TEST_SOURCE_SET_NAME)
-  @Classpath final def getRuntimeClassPath: FileCollection = sourceSet.getRuntimeClasspath
-
-  // To avoid invoking Task.getProject at execution time, some things are done or captured in afterEvaluate():
-  private var analysisFile: Option[File] = None
-
-  getProject.afterEvaluate((project: Project) =>
-    getDependsOn.add(project.getClassesTask(sourceSet))
-
-    // AnalysisDetector needs Zinc classes;
-    // if I ever get rid of it, this classpath expansion goes away.
-    addToClassPath(this, project.getConfiguration(ScalaBasePlugin.ZINC_CONFIGURATION_NAME).asScala)
-
-    // Note: scalaCompile.getAnalysisFiles is empty, so I had to hard-code the path:
-    analysisFile = Some(Files.file(
-      directory = project.getLayout.getBuildDirectory.get.getAsFile,
-      segments = s"tmp/scala/compilerAnalysis/${project.getScalaCompile(sourceSet).getName}.analysis"
-    ))
-  )
+  final override protected def sourceSetName: String = SourceSet.TEST_SOURCE_SET_NAME
 
   final def useSbt(@DelegatesTo(classOf[TestFrameworkOptions]) testFrameworkConfigure: Closure[?]): Unit =
     useSbt(ConfigureUtil.configureUsing(testFrameworkConfigure))
@@ -90,6 +78,16 @@ abstract class TestTask extends Test:
     // best be done with this here, before `super.createTestExecuter()` is called.
     require(getTestFramework.isInstanceOf[TestFramework], s"Only useSbt Gradle test framework is supported by this plugin - not $testFramework!")
     require(isScanForTestClasses, "File-name based test scan is not supported by this plugin, `isScanForTestClasses` must be `true`!")
+
+    // AnalysisDetector needs Zinc classes;
+    // if I ever get rid of it, this classpath expansion goes away.
+    // Note: when done at creation, causes:
+    //   Cannot change dependencies of dependency configuration ':zinc' after it has been resolved.
+    GradleClassPath.addTo(this, zincClassPath)
+
+    // Note: scalaCompile.getAnalysisFiles is empty, so I had to hard-code the path:
+    analysisFile = Some(Files.file(buildDirectory, s"tmp/scala/compilerAnalysis/$scalaCompile.analysis"))
+
     super.executeTests()
 
   final override def createTestExecuter: TestExecuter = TestExecuter(